% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_world_map_custom_center.R
\name{create_ocean_background}
\alias{create_ocean_background}
\title{object ocean as an oval for prime meridian at custom longitude}
\usage{
create_ocean_background(long)
}
\arguments{
\item{long}{longitude of the custom center}
}
\value{
sf object as oval polygon
}
\description{
object ocean as an oval for prime meridian at custom longitude
}
