% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_world_map_custom_center.R
\name{create_world_map_custom_center}
\alias{create_world_map_custom_center}
\title{create world map as sf object}
\usage{
create_world_map_custom_center(center = 100, class = "sf")
}
\arguments{
\item{name}{description}
}
\value{
sf object
}
\description{
create world map as sf object
}
\examples{
sf = create_world_map_custom_center()
ggplot() + geom_sf(data =sf)
}
