% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentize_polygon.R
\name{segmentize_polygon}
\alias{segmentize_polygon}
\title{segmentize the polygon so that we can have smoother line after projection}
\usage{
segmentize_polygon(sf, crs = myCrs)
}
\arguments{
\item{sf}{sf of the polygon to smooth out}
}
\value{
sf segmentized on the slice side
}
\description{
The polygon need to be cast into POLYGON for the shape to not get distorted
}
