% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftCenterWorldMap-package.R
\docType{package}
\name{shiftCenterWorldMap-package}
\alias{shiftCenterWorldMap}
\alias{shiftCenterWorldMap-package}
\title{shiftCenterWorldMap: shift the center of world map plot in ggplot in robinson projection}
\description{
generate world map sf object vignette on animation
}
\seealso{
Useful links:
\itemize{
  \item \url{https://thehung92.github.io/shiftCenterWorldMap/}
}

}
\keyword{internal}
