% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftCenterWorldMap-package.R
\docType{package}
\name{shiftCenterWorldMap-package}
\alias{shiftCenterWorldMap}
\alias{shiftCenterWorldMap-package}
\title{shiftCenterWorldMap: shift the center of world map plot in ggplot in robinson projection}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
